// hottub-web/src/generated/model/settings.ts
/**
 * Hottub API
 * API for hottub
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Settings { 
    /**
     * Wanted temperature in hot tub
     */
    temperatureLimit: number;
    /**
     * Delta temperature difference between sensed temperature and actual temperature. Needed since temperature sensor is in the bottom, meaning temperatures will differ a few degrees from the sensor and the top of the hottub
     */
    temperatureDelta: number;
    /**
     * Maximum temperature allowed in heating pan until circulation pump should start operating
     */
    heatingPanTemperatureLimit: number;
    /**
     * Number of minutes between circulation pump operations, to make sure nothing freezes in winter and to stir around the water a bit
     */
    circulationTimeCycle: number;
    /**
     * Whether the lights should be on or off
     */
    lightsOn: boolean;
    /**
     * Activate debug logs, which will log a lot more of operational information
     */
    debugMode: boolean;
}