// hottub-web/src/generated/model/timer.ts
/**
 * Hottub API
 * API for hottub
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Settings } from './settings';


export interface Timer { 
    settings: Settings;
    /**
     * Timestamp for when this settings should kick in
     */
    time: string;
}