// hottub-web/src/generated/model/timerEntity.ts
/**
 * Hottub API
 * API for hottub
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Timer } from './timer';
import { TimerEntityAllOf } from './timerEntityAllOf';
import { Settings } from './settings';


/**
 * An existing timer entity
 */
export interface TimerEntity { 
    settings: Settings;
    /**
     * Timestamp for when this settings should kick in
     */
    time: string;
    id: string;
}